requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-telegram-bot==20.7
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
